Add mongo source and note model
