+++ src/datasource/database.module.ts
@@ -1,20 +1,8 @@
-import * as mongoose from 'mongoose';
+import { Module } from '@nestjs/common';
+import { databaseProviders } from './database.providers';
 
-
-// Envirionment variabled
-const USERNAME = process.env.USERNAME;
-const PASSWORD = process.env.PASSWORD;
-const DB_NAME = process.env.DB_NAME;
-
-
-// connection string to mongo atlas
-const uri = `mongodb+srv://${USERNAME}:${PASSWORD}@lenda.padmy06.mongodb.net/?retryWrites=true&w=majority&appName=${DB_NAME}`;
-
-
-export const databaseProviders = [
-  {
-    provide: 'DATABASE_CONNECTION',
-    useFactory: (): Promise<typeof mongoose> =>
-      mongoose.connect('mongodb://localhost/nest'),
-  },
-];
+@Module({
+  providers: [...databaseProviders],
+  exports: [...databaseProviders],
+})
+export class DatabaseModule {}
